metasejong-server/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.35
pymysql==1.1.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
